requirements.txt
paddlepaddle==2.6.2
paddleocr==2.7.3
numpy<2.0.0
protobuf==3.20.2
piexif
Pillow
tqdm
